+++ messaging/messaging-jms/src/main/java/uk/gov/justice/services/messaging/jms/JmsEnvelopeSenderProducer.java
@@ -1,5 +1,7 @@
 package uk.gov.justice.services.messaging.jms;
 
+import static uk.gov.justice.services.core.annotation.Component.COMMAND_CONTROLLER;
+
 import uk.gov.justice.services.common.annotation.ComponentNameExtractor;
 
 import javax.enterprise.inject.Produces;
@@ -20,10 +22,14 @@ public class JmsEnvelopeSenderProducer {
     @Produces
     public JmsEnvelopeSender createJmsEnvelopeSender(final InjectionPoint injectionPoint) {
 
-        if (componentNameExtractor.hasComponentAnnotation(injectionPoint)) {
+        if (componentNameExtractor.hasComponentAnnotation(injectionPoint) && !isCommandController(injectionPoint)) {
             return new ShutteringJmsEnvelopeSender(envelopeSenderSelector);
         }
 
         return new DefaultJmsEnvelopeSender(jmsSender);
+    }
+
+    private boolean isCommandController(final InjectionPoint injectionPoint) {
+        return COMMAND_CONTROLLER.equals(componentNameExtractor.componentFrom(injectionPoint));
     }
 }