+++ CHANGELOG.md
@@ -8,6 +8,7 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 ## [6.1.2] - 2018-10-04
 ### Changed
 - Storing of Commands whilst shuttered is now idempotent 
+- Do not shutter the JmsEnvelopeSender when called from the Command-Controller component
 
 ## [6.1.1] - 2018-10-01
 ### Added